feat: add 'forKit' field to services and update related components
- Added 'forKit' field to the Service model and updated the database schema.
- Modified service creation and retrieval logic to handle 'forKit' values.
- Updated the CreateUpdateService component to include a dropdown for selecting 'forKit'.
- Enhanced the Services component to display total price based on selected services.
- Updated ItemSelection and KitSelection components to manage total price calculations.
- Added styles for total price display in relevant components.
- Adjusted API routes and service controller to accommodate new 'forKit' functionality.

+++ web/app/lib/services/service.service.ts
@@ -3,6 +3,7 @@ import { put, del } from "@vercel/blob";
 import { CreateService, EditService } from "../utils/requests/service.request";
 import { ServiceResponses } from "../utils/responses/serviceResponses.";
 import { AppError } from "../withError";
+import { KitType } from "@prisma/client";
 
 type FileOrString = File | string;
 
@@ -12,10 +13,11 @@ export type ServiceSearchPayload = {
 
 export const ServiceService = {
   async create(formData: FormData) {
-    const { name, icon, price }: CreateService = {
+    const { name, icon, price, forKit }: CreateService = {
       name: String(formData.get("name")),
       price: Number(formData.get("price")),
-      icon: formData.get("icon") as File
+      icon: formData.get("icon") as File,
+      forKit: formData.get("forKit") as KitType
     }
 
     const serviceId = crypto.randomUUID();
@@ -26,13 +28,16 @@ export const ServiceService = {
       { access: "public" }
     );
 
+    const forKitType = Object.values(KitType).includes(forKit) ? forKit : "ALL";
+
     try {
       return await prisma.service.create({
         data: {
           id: serviceId,
           name: name.trim().normalize("NFC").toLowerCase(),
           price: price,
-          icon: blob.url
+          icon: blob.url,
+          forKit: forKitType
         }
       });
 
@@ -65,16 +70,34 @@ export const ServiceService = {
     return services;
   },
 
-  async getAll() {
-    return await prisma.service.findMany();
+  async getAll(kitType: KitType) {
+    const param = Object.values(KitType).includes(kitType) ? kitType : "ALL";
+    const filter = [param];
+
+    if (filter.includes("ALL")) {
+      return await prisma.service.findMany();
+    }
+
+    filter.push("ALL");
+
+    return await prisma.service.findMany({
+      where: {
+        forKit: {
+          in: filter
+        }
+      }
+    });
   },
 
   async edit(id: string, formData: FormData) {
-    const { name, icon, price }: EditService = {
+    const { name, icon, price, forKit }: EditService = {
       name: String(formData.get("name")),
       price: Number(formData.get("price")),
       icon: formData.get("icon") as FileOrString,
+      forKit: formData.get("forKit") as KitType
     }
+
+    const forKitType = Object.values(KitType).includes(forKit) ? forKit : "ALL";
 
     try {
       const currentService = await prisma.service.findUnique({
@@ -104,7 +127,8 @@ export const ServiceService = {
 
       if (
         currentService.name !== name ||
-        !currentService.price.equals(price)
+        !currentService.price.equals(price) ||
+        currentService.forKit !== forKit
       ) {
         await prisma.service.update({
           where: {
@@ -113,7 +137,8 @@ export const ServiceService = {
           data: {
             name: name.trim().normalize("NFC").toLowerCase(),
             price: price,
-            icon: image
+            icon: image,
+            forKit: forKitType
           }
         });
 
@@ -134,6 +159,14 @@ export const ServiceService = {
 
   async delete(id: string) {
     try {
+      const service = await prisma.service.findUnique({
+        where: { id }
+      });
+
+      if (!service) throw new AppError(404, ServiceResponses.SERVICE_NOT_FOUND);
+
+      await del(service.icon);
+
       await prisma.service.delete({
         where: {
           id
@@ -141,7 +174,7 @@ export const ServiceService = {
       });
 
     } catch {
-      throw new AppError(404, ServiceResponses.SERVICE_DELETED_ERROR);
+      throw new AppError(400, ServiceResponses.SERVICE_DELETED_ERROR);
     }
   },
